refactor: create getSkillStats
